backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langgraph>=0.2.0
langchain>=0.2.16
langchain-openai>=0.1.7
langchain-community>=0.2.16
langchain-text-splitters>=0.2.4
langchain-core>=0.2.43
chromadb>=0.5.0
pydantic>=2.7.4
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
